% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertSolution.R
\name{ConvertSolution}
\alias{ConvertSolution}
\title{Converts lines containing certain strings to other lines}
\usage{
ConvertSolution(x, answerstring = "# Answer",
  hidestring = "____add answer___", stopstring = "# Stop and think: ",
  collapsemany = TRUE)
}
\arguments{
\item{x}{The vector of input strings (representing one line each)}

\item{answerstring}{The tag indicating lines to convert to hide}

\item{hidestring}{What to replace the answer lines with}

\item{stopstring}{The tag indicating what to prompt user to do}

\item{collapsemany}{If you have multiple answer strings in order, collapse}
}
\value{
The vector of transformed strings
}
\description{
When I teach, I like to have a completed solution made first (so I know it
works) and then modify this for students to fill in. It could be simple
matters of filling in a function name or supplying an argument
or it could be writing a complex series of functions. This requires manual
deletion or changing of lines each time the solution file is changed.
With this function, anything with the answerstring in it will be converted to
the hidestring in the output (which could then be written to an exercise
file). I sometimes insert stop functions, as well, to keep students from
charging forward without thinking -- they have to do what it says at this
this point. This can also be included.
}
\examples{
input <- c("# Count the number of tips", "phy <- ape::rcoal(15)",
"print(Ntip(phy)) # Answer",
"# Stop and think: How can you know this answer is correct?")
output <- ConvertSolution(input)
print(output)
}

